Forge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forge-cli"
version = "1.0.0"
description = "Local-first CLI orchestration tool for fine-tuning Small Language Models using Gemini 3"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Forge Team"}
]
keywords = ["llm", "fine-tuning", "gemini", "cli", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "keyring>=25.0.0",
    "psutil>=5.9.0",
    "nvidia-ml-py>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
]
# Heavy ML dependencies - only needed if running training locally (not via Docker)
local-training = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.28.0",
]
unsloth = [
    "unsloth[colab-new]>=2024.4",
]

[project.scripts]
forge = "forge.cli.main:app"

[project.urls]
Homepage = "https://github.com/forge-cli/forge"
Documentation = "https://forge-cli.github.io"
Repository = "https://github.com/forge-cli/forge"

[tool.hatch.build.targets.wheel]
packages = ["forge"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
